requirements/requirements.txt
Django==4.0.4
django-braces==1.15.0
django-filter==22.1
django-localflavor==3.1
django-redis==5.2.0
uuid==1.30
requests==2.28.1
unicode==2.9
wkhtmltopdf
Pillow
matplotlib
numpy
psycopg2-binary

aldjemy==2.6
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
awscli==1.25.25
botocore==1.27.25
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.4
cryptography==37.0.4
Deprecated==1.2.13
distro==1.5.0
docutils==0.16
escape==1.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
Jinja2==2.10
jmespath==1.0.1
json5==0.9.8
markup==0.2
MarkupSafe==2.0.1
MouseInfo==0.1.3
outcome==1.2.0
packaging==21.3
pdfkit==1.0.0
pyasn1==0.4.8
pycparser==2.21
PyGetWindow==0.0.9
pymemcache==3.5.2
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-stdnum==1.17
python3-xlib==0.15
pytweening==1.0.4
pyxdg==0.27
PyYAML==5.4.1
redis==4.3.4
rsa==4.7.2
s3transfer==0.6.0
scape==0.9.9
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.10
wrapt==1.14.1
wsproto==1.1.0
xdg==5